Finished Exercise 3: JAX-RS

+++ service/src/main/java/es/uvigo/esei/dgss/exercises/service/sample/PostEJB.java
@@ -4,14 +4,19 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+import javax.annotation.Resource;
+import javax.ejb.EJB;
+import javax.ejb.SessionContext;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
 import es.uvigo.esei.dgss.exercises.domain.Comment;
+import es.uvigo.esei.dgss.exercises.domain.Link;
 import es.uvigo.esei.dgss.exercises.domain.Photo;
 import es.uvigo.esei.dgss.exercises.domain.Post;
 import es.uvigo.esei.dgss.exercises.domain.User;
+import es.uvigo.esei.dgss.exercises.domain.Video;
 
 @Stateless
 public class PostEJB {
@@ -19,18 +24,57 @@ public class PostEJB {
 	@PersistenceContext
 	private EntityManager em;
 
+	@EJB
+	private UserEJB userejb;
+
+	@EJB
+	private StatisticsEJB stats;
+
+	@Resource
+	private SessionContext ctx;
+
+	private void checkUserAccess(String login, String message) {
+		if (!ctx.getCallerPrincipal().getName().equals(login) && !ctx.isCallerInRole("admin")) {
+			throw new SecurityException(message);
+		}
+	}
+
 	public Post findPostById(int id) {
 		return em.find(Post.class, id);
 	}
 
-	public Post addPost(User user, String content) {
-		Post post = new Photo(user, content);
+	public Link addLink(Link link) {
+		checkUserAccess(link.getPoster().getLogin(), "Security Error: You cannot add links on behalf of other users.");
 
-		em.persist(post);
-		return post;
+		this.em.persist(link);
+		stats.addPost();
+
+		return link;
+	}
+
+	public Photo addPhoto(Photo photo) {
+		checkUserAccess(photo.getPoster().getLogin(),
+				"Security Error: You cannot add photos on behalf of other users.");
+
+		this.em.persist(photo);
+		stats.addPost();
+
+		return photo;
+	}
+
+	public Video addVideo(Video video) {
+		checkUserAccess(video.getPoster().getLogin(),
+				"Security Error: You cannot add videos on behalf of other users.");
+
+		this.em.persist(video);
+		stats.addPost();
+
+		return video;
 	}
 
 	public void updatePost(Post post) {
+		checkUserAccess(post.getPoster().getLogin(), "Security Error: You cannot modify posts from other users.");
+
 		Post updated = findPostById(post.getId());
 		updated.setComments(post.getComments());
 		updated.setDate(post.getDate());
@@ -39,45 +83,42 @@ public class PostEJB {
 	}
 
 	public void removePost(int id) {
+		checkUserAccess(findPostById(id).getPoster().getLogin(),
+				"Security Error: You cannot remove posts from other users.");
+
 		em.remove(findPostById(id));
 	}
 
 	@SuppressWarnings("unchecked")
-	private List<Post> getPosts(User user) {
+	public List<Post> getPosts(User user) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot access posts from other users.");
+
 		return (List<Post>) em.createQuery("SELECT p FROM Post p WHERE p.poster = :user").setParameter("user", user)
 				.getResultList();
-	}
-
-	@SuppressWarnings("unchecked")
-	public List<User> getFriends(User user) {
-		List<User> friends = (List<User>) em.createQuery("SELECT f.friend2 FROM Friend f WHERE f.friend1 = :user")
-				.setParameter("user", user).getResultList();
-
-		List<User> befriendedBy = (List<User>) em.createQuery("SELECT f.friend1 FROM Friend f WHERE f.friend2 = :user")
-				.setParameter("user", user).getResultList();
-
-		friends.addAll(befriendedBy);
-		return friends;
 	}
 
 	public List<Post> getFriendPosts(User user) {
 		List<Post> posts = new ArrayList<Post>();
 
-		getFriends(user).forEach(f -> posts.addAll(getPosts(f)));
+		userejb.getFriends(user).forEach(f -> posts.addAll(getPosts(f)));
 
 		return posts;
 	}
 
 	@SuppressWarnings("unchecked")
 	private List<Comment> getComments(User user) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot access comments from other users.");
+
 		return (List<Comment>) em.createQuery("SELECT c FROM Comment c WHERE c.user = :user").setParameter("user", user)
 				.getResultList();
 	}
 
 	public List<Post> getPostsCommentedAfterDate(User user, Date date) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot access comments from other users.");
+
 		List<Post> posts = new ArrayList<Post>();
 
-		getFriends(user).forEach(f -> getComments(f).forEach(c -> {
+		userejb.getFriends(user).forEach(f -> getComments(f).forEach(c -> {
 			if (c.getDate().after(date))
 				posts.add(c.getPost());
 		}));
@@ -87,6 +128,8 @@ public class PostEJB {
 
 	@SuppressWarnings("unchecked")
 	private List<Post> getLiked(User user) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot access likes from other users.");
+
 		return (List<Post>) em.createQuery("SELECT u.likes FROM User u WHERE u.login = :user")
 				.setParameter("user", user.getLogin()).getResultList();
 	}
@@ -94,7 +137,7 @@ public class PostEJB {
 	public List<User> getFriendsWhoLikedPost(User user, Post post) {
 		List<User> users = new ArrayList<User>();
 
-		getFriends(user).forEach(f -> {
+		userejb.getFriends(user).forEach(f -> {
 			if (getLiked(f).contains(post))
 				users.add(f);
 		});
