+++ service/src/main/java/es/uvigo/esei/dgss/exercises/service/sample/UserEJB.java
@@ -1,7 +1,6 @@
 package es.uvigo.esei.dgss.exercises.service.sample;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import javax.annotation.Resource;
@@ -22,10 +21,22 @@ public class UserEJB {
 	private EntityManager em;
 
 	@EJB
+	private PostEJB postejb;
+
+	@EJB
+	private StatisticsEJB statsejb;
+
+	@EJB
 	private EmailServiceEJB emailejb;
 
 	@Resource
 	private SessionContext ctx;
+
+	private void checkUserAccess(String login, String message) {
+		if (!ctx.getCallerPrincipal().getName().equals(login) && !ctx.isCallerInRole("admin")) {
+			throw new SecurityException(message);
+		}
+	}
 
 	public User findUserById(String login) {
 		return em.find(User.class, login);
@@ -39,14 +50,20 @@ public class UserEJB {
 		user.setPicture(picture);
 
 		em.persist(user);
+		statsejb.addUser();
+
 		return user;
 	}
 
-	public void updateUser(User user) {
+	public User updateUser(User user) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot modify data from other users.");
+
 		User updated = em.find(User.class, user.getLogin());
 		updated.setName(user.getName());
 		updated.setPassword(user.getPassword());
 		updated.setPicture(user.getPicture());
+
+		return updated;
 	}
 
 	private void removeRelationships(User user) {
@@ -57,28 +74,75 @@ public class UserEJB {
 	}
 
 	public void removeUser(String login) {
+		checkUserAccess(login, "Security Error: You cannot remove data from other users.");
+
 		removeRelationships(em.find(User.class, login));
 		em.createQuery("DELETE FROM User u WHERE u.login = :login").setParameter("login", login).executeUpdate();
+		statsejb.removeUser();
+	}
+
+	public List<Post> getWall(User user) {
+		List<Post> wall = this.postejb.getPosts(user);
+		wall.addAll(this.postejb.getFriendPosts(user));
+
+		return wall;
 	}
 
 	public void addLike(User user, Post post) {
-		List<Post> likedPosts = user.getLikes();
-		likedPosts.add(post);
-		user.setLikes(likedPosts);
+		checkUserAccess(user.getLogin(), "Security Error: You cannot like posts on behalf of other users.");
 
-		List<User> postLikers = post.getLikers();
-		postLikers.add(user);
-		post.setLikers(postLikers);
+		if (this.getWall(user).contains(post)) {
+			List<Post> likedPosts = user.getLikes();
+			likedPosts.add(post);
+			user.setLikes(likedPosts);
 
-		emailejb.sendEmail(post.getPoster(), "New like", user.getName() + " has liked your post");
+			List<User> postLikers = post.getLikers();
+			postLikers.add(user);
+			post.setLikers(postLikers);
+
+			emailejb.sendEmail(post.getPoster(), "New like", user.getName() + " has liked your post");
+		} else {
+			throw new SecurityException("You can only like posts from your wall.");
+		}
+	}
+
+	public void removeLike(User user, Post post) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot remove likes from other users.");
+
+		if (this.getWall(user).contains(post)) {
+			List<Post> likedPosts = user.getLikes();
+			likedPosts.remove(post);
+			user.setLikes(likedPosts);
+
+			List<User> postLikers = post.getLikers();
+			postLikers.remove(user);
+			post.setLikers(postLikers);
+		} else {
+			throw new SecurityException("You can only remove likes from your wall.");
+		}
 	}
 
 	public Friend addFriendship(User user1, User user2) {
 		Friend friendship = new Friend(user1, user2);
-		friendship.setDate(new Date());
 
 		em.persist(friendship);
 		return friendship;
+	}
+
+	public void acceptFriendship(User user1, User user2) {
+		em.createQuery("UPDATE Friend f SET accepted = TRUE WHERE f.friend1 = :user2 AND f.friend2 = :user1")
+				.setParameter("user1", user1).setParameter("user2", user2).executeUpdate();
+	}
+
+	public void rejectFriendship(User user1, User user2) {
+		em.createQuery("DELETE FROM Friend f WHERE f.friend1 = :user2 AND f.friend2 = :user1")
+				.setParameter("user1", user1).setParameter("user2", user2).executeUpdate();
+	}
+
+	@SuppressWarnings("unchecked")
+	public List<User> getFriendRequests(final User user) {
+		return (List<User>) em.createQuery("SELECT f.friend1 FROM Friend f WHERE f.friend2 = :user AND accept = FALSE")
+				.setParameter("user", user).getResultList();
 	}
 
 	@SuppressWarnings("unchecked")
@@ -94,6 +158,8 @@ public class UserEJB {
 	}
 
 	public List<User> getPotentialFriends(User user) {
+		checkUserAccess(user.getLogin(), "Security Error: You cannot access data from other users.");
+
 		List<User> friends = new ArrayList<User>();
 		getFriends(user).forEach(u -> friends.addAll(getFriends(u)));
